+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - Added more icons and changed icons to come from HDS.
 - Added Search Block.
 - Added React & Share Block.
-- Added Tip style for Paragraph.
+- Added Tip and Info styles for Paragraph.
 - Added captions for page and post featured images.
 - Added new header for posts and pages.
 - Added sidebar to posts.